Basic seek support
For now just mimic the previous behaviour of babeltrace but with a basic
seek structure.
We define the babeltrace_iter_seek that seeks each file_stream of an
iterator to a given position and the babeltrace_filestream_seek function
that does the actual seek on the file_stream and can be used internally.

Signed-off-by: Julien Desfossez <julien.desfossez@polymtl.ca>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>