Add mmap trace reading
This set of functions allows a client to open a trace that is not stored
on tracefiles but instead is in memory.
Since the parameters required for such an operation are really different
than typical traces, we define the open_mmap_trace function.
This patch also introduces the mmap_stream structure and
mmap_stream_list that represents a list of FDs readable by mmap.

[ Edit by Mathieu Desnoyers: cleanup ]

Signed-off-by: Julien Desfossez <julien.desfossez@polymtl.ca>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>