Fix parallel build for lib/
When we have a dependency from a library on another library (with
LIBADD), we need to put the lib we depend on in another directory, and
build that directory prior to the current directory. This is what is
done in the http://sourceware.org/autobook/autobook/autobook_92.html
"11.5 Convenience Libraries" example.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>