Introduce contexts
The context is an object in which the trace_collection is opened. The
iterator (and later callbacks) depend on the context. This patch, does
not modify the behaviour, it is just the introduction of the contexts.
Also moved convert_trace() in the converter code since it is not a
public API function.

Signed-off-by: Julien Desfossez <julien.desfossez@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>