Revert "Remove leading underscores from identifiers directly in lexer"
This reverts commit f071e095a2a8f76dd78becfa103b952af68e148d.

The reason why the underscores are needed in the first place is because
we want to make sure sequences and variants to not have clash between
fields called e.g. "stream" and the dynamic scope
"stream.event.header.". Therefore, we need to keep the leading
underscores as part of the internal identifier representation.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>