Fix : segfault when printing timestamp on index
If we detect we lost events in a packet, we print an error message
containing the begin and end timestamps of the packet.  If we detect
this packet loss while generating the index (during the open_trace), the
trace_collection is not yet referenced and so the clock information is
not available. This fix prevents the printing at this stage.

Signed-off-by: Julien Desfossez <julien.desfossez@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>