Fix: seek by timestamp
When seeking a stream by timestamp, we need to match the first packet in
the stream that is at (or after) the timestamp to seek to. Failure to do
so is a bug: if we have many streams, and seek all streams to a
timestamp that is after the start time of some of the streams, those
late comers will show no events.

Reported-by: Yannick Brosseau <yannick.brosseau@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>