Fix error checking in bt_context_add_traces_recursive
If a trace directory is inexistant or if the user doesn't have the
permission to open the parent directory, bt_context_add_traces_recursive
returned 0, now it returns -1. This function is not part of the API, so
the programs who copied this function into their code should apply this
fix as well.

Fixes #103

Signed-off-by: Julien Desfossez <julien.desfossez@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>