API Fix : missing list fields of event decl
This patchs allows the user to list the fields (in all scopes) for each
event declarations (so as soon as the trace is opened).
It uses a lazy allocation, so the memory for storing the fields is only
allocated if the users asks for the list and it is allocated per event.

Signed-off-by: Julien Desfossez <julien.desfossez@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>