Fix: --clock-force-correlate to handle trace collections gathered from various nodes
Currently, traces generated by lttng have the "absolute" flag set to 0,
because they are based on the monotonic clock, which is not perfectly
correlated across nodes.

Fix a couple of minor issue in the handling of --clock-force-correlate
that now allow to use it to force correlation of traces taken with the
monotonic clock source, with "offset" approximation populated using NTP
time.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>