Add new option --clock-offset-ns
Add a new option --clock-offset-ns. It requires a parameter
that specifies a clock offset (in nanoseconds) that is added
to each timestamp.

This works exactly as the option --clock-offset, except that
the value is given in nanoseconds instead of full seconds.

The two options --clock-offset and --clock-offset-ns are
compatible with each other, and it's possible to give both.
For example, having the options

--clock-offset 2 --clock-offset-ns 1000000

means that 2.001 seconds is added to each timestamp.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>