Add MinGW implementation of UUID functions
Add implementation of the UUID functions for MinGW.
The Windows UUID standard has a different endian from what we expect,
so also do endian conversion for the UUID data types as needed.
Add the MinGW implementation file to the build system.
In configure.ac, assume that MinGW UUID functions exist.

[ Edit by Mathieu Desnoyers: coding style fixes, use strncmp rather than
  strcmp. Use already existing defines rather than hardcoded integers. ]

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>