Extract adding a trace to the iterator from bt_iter_init
With live tracing, we have cases where we need to add traces to an
iterator currently in use. Here, we just extract the code from
bt_iter_init that adds the trace to the iterator so we can call it from
elsewhere.
We don't necessarily want to expose this interface to the public, so it
stays in the internal headers for now.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>