Attach and list by session name and hostname
From now on, the user attaches to a session by hostname and session name
(no ID anymore). With this new features, we automatically merge the
sessions with the same hostname and session name.

The new URL format to attach to a session is :
net://<relay-hostname>/host/<traced-hostname>/<session-name>

All the fields are mandatory.
No changes for the listing :
net://<relay-hostname>

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>