Python: decouple Python iface from native binding
With this patch, all the SWIG stuff is isolated into
nativebt.i, whereas the real Babeltrace Python interface
is plain Python in bt.py.

The Babeltrace Python binding is now a Python package, but
backward compatibility with the original module is ensured
by importing all public symbols from the bt module in the
package itself (__init__.py), which is named like the
original module, babeltrace.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>