Allow unset tags on variant field types
This allows the creation of a variant while not specifying the variant's
tag's name and type which is needed to support untagged variants.

This will allow the support of constructs such as:
variant name {
    field_type sel1;
    field_type sel2;
    field_type sel3;
    /* ... */
};

struct {
    enum : integer_type { sel1, sel2, sel3, /* ... */ } tag_field;
    /* ... */
    variant name <tag_field> v;
}

The validation of variant fields thus becomes optional and will
be implemented as part of the freeze() of variants in a later commit.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>