Fix: Ensure stream endianness is set when initializing event class
This fixes a problem encountered by perf-to-ctf which would cause
an assert to hit whenever an event was instanciated before its
stream class was added to the trace.

This would cause the stream class to set its events' "native"
byte order with the "unknown endianness" value.

This modification ensures that an event's native endianness is
either set on addition to a stream class (when the stream is
already paired to a trace) or when the stream is added to a
trace.

Reported-by: Wang Nan <wangnan0@huawei.com>
Reported-by: Arnaldo Carvalho de Melo <acme@redhat.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>