objects: add bt_object_freeze() and more statuses
This patch adds the possibility to freeze an object.
By doing this, it introduces a new status code,
BT_OBJECT_STATUS_FROZEN, which is returned when trying
to alter a frozen object. Thus, all functions returning
either 0 or -1 are modified to return an
enum bt_object_status value instead.

The status code BT_OBJECT_STATUS_INVAL is also introduced
to signal invalid arguments, to differentiate this error
from other errors.

Finally, this patch adapts the objects tests and the
documentation to the changes.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>