Fix: Remove overly-strict freeze check on variants and structures
bt_ctf_field_type_structure_set_field_index() and
bt_ctf_field_type_variant_set_field_index() are both internal APIs
which are used when resolving field paths.

For exemple, when a sequence's field path is resolved, the sequence
is copied and replaced in its parent since we are setting the
resolved path as part of the sequence. This is done so that different
sequence members which would have been declared using the same sequence
type instance can have length fields which resolve to different absolute
paths.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>