Fix: do not init. BO of enums/arrays/sequences
The functions used to create enumeration, array, and
sequence types are different from the other creation
functions in that they register another type at creation
time. Other compound types are empty when created. When a
type is initialized, its byte order setting function is
called, which is recursive for compound types and for the
enumeration type. This overrides the previously set byte
order of the underlying type passed to those creation functions.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>