Fix: ir: visitor.c: handle type copies correctly
A copy is created in type_resolve_func(), and it replaces
the original field type in its compound type, but this
copy is not the one pushed onto the stack afterwards; the
original, weak reference is pushed.

The current code of visitor.c also has an issue where
field_type_recursive_visit() gets a reference of a
field type, field_type_visit() pushes this reference
onto the stack, and then field_type_recursive_visit()
puts this reference. This is fixed by getting a
reference when pushing onto the stack and putting it
when popping it.

The recursion happening when array and sequence types
are encountered makes this code somewhat complex for
reference management. This patch modifies this to also
push array and sequence types onto the stack, since they
are compound types too, therefore removing recursion
completely. It also makes sure to skip array and sequence
types present in the stack when composing a field path.

Finally, this patch also makes sure to update the array
and sequence types element types with a copy when they
contain sequence or variant types.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>