Python bindings: work around Python 3.5 behaviour change
Python 3.5 changes the StopIteration exception clearing behaviour
when a generator finishes its iteration. This causes the
interpreter to errounously consider SWIG clean-up functions as
having "set an error".

This hack explicitly allocates and cleans up struct bt_iter_pos
instead of relying on SWIG auto-generated code which manages
the lifetime of temporary objects.

An investigation of the cause of this change is under way, but
at least this makes the bindings usable on Python 3.5 which is
being rolled-out in some distros.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>