ir: move the stream event ctx field to the event
Just like the event header, the stream event context field has
its type defined in the stream class, but its data can be different
in each event. The current strategy to sample and copy the current
stream's stream event context and append the copies to an array
is fine as far as CTF writer goes, but this array would not be
populated strictly using the event in the context of Babeltrace
components.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>