Fix: Make test_dwarf and test_bin_info tests work out-of-tree
When building out of tree, these tests fail to find the test data.  This
patch makes them (their _complete wrappers, actually) generated by
Autoconf, so the proper path to the test data (which is in the source
directory) can be written.

We can also get rid of SCRIPT_LIST and it's custom rules.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>