Change behaviour of stream-intersection with multiple traces
The stream-intersection currently results in the reading of
the intersection of every stream, in every trace of a Trace
Collection.

Since reading a TraceCollection returns the union of its traces,
the behaviour of this option is changed so that it returns the
union of each trace's individual "active" section, that is the
section for which all streams are active.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>