python: Add stream event context support
This patch adds the possibility to get/set the event context type of a
stream class and set the stream context value of an event from Python.

From my experience, the value setter (Event.stream_context.setter) is
not so useful, since it looks like the context is already instanciated
when you create the event.  So in order to set the values in it, you
only need to get the existing instance and set its fields.  There is no
need to to re-assign the stream context instance to the event.  I still
put it there for completeness.

I added an example of the usage to the existing CTF writer Python
example.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>