Text output plugin argument parsing
Implement plugin argument handling, and legacy behavior.

The sink plugin uses "name-default=show", "name-default=hide",
"field-default=show", and "field-default-hide" to specify the default
behavior for names and fields. It is used to implement the compability
layer with Babeltrace 1.x (all/none), but is a more flexible approach,
because it then allows both use-cases:

- default=show: can then set individual items to "false" (blacklist),
- default=hide: can then set individual items to "true" (whitelist).

In comparison, the old legacy options only allowed to specify a
whitelist.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>