m4/ax_lib_elfutils.m4: use m4_default() for optional args.
The current macro code does not work because, if you provide 4
arguments to the macro, then `true_action` is not defined (but
it should be: it's the third argument).

Use m4_default() here to expand to a default value if the
argument does not exist.

Also use [:] instead of [] (good practice, safer for shells).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>