babeltrace-cfg: add --base-params and --reset-base-params options
The new --base-params option sets the current "base parameters" of
the command-line parsing context. Any --source or --sink which comes
after this option is initialized with those parameters. The --params
now has the effect of extending (patching) the current instance
parameters.

The base parameters can be changed at any place in the command line
with the --base-params option (can be repeated).

The new --reset-base-params option resets the base parameters to
an empty map where it is placed in the command line (can be
repeated).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>