Add bt_component_create_with_init_method_data()
This function has one more parameter than bt_component_create():
custom user data (void *) which is passed directly to the
init. method (if any) of the used component class. Nothing else
is done by the lib with this data.

bt_component_create() is a specific version of
bt_component_create_with_init_method_data() which passes NULL
as the init. method data.

You can use this custom data to exchange specific data between
an application and a component class. It is not recommended to
rely on this function when you implement Babeltrace plugins.

A component initialization method now has an additional parameter,
init_method_data.

This initialization method data is not related at all to a
component's private data: you can set the init. method data as
the component's private data, but it's just a design choice.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>